refactor furioso
