+++ server/router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 from .models import Challenge
-from server.game import Game
-from server.server import games, notify_challenge_to_client
+from server.game import Game, games
+import server.websockets
 
 router = APIRouter()
 
@@ -10,7 +10,7 @@ router = APIRouter()
 async def challenge(challenge: Challenge):
     game = Game(challenge.challenger, challenge.challenged)
     games.append(game)
-    await notify_challenge_to_client(
+    await server.websockets.notify_challenge_to_client(
         game.player,
         game.challenged_player,
         game.uuid_game,
